requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
argparse==1.4.0
backoff==2.2.1
beautifulsoup4==4.13.3
browser-use==0.1.40
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
deprecated==1.2.18
distro==1.9.0
filelock==3.18.0
filetype==1.2.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
importlib-metadata==8.6.1
inflection==0.5.1
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==0.3.3
langchain-core==0.3.50
langchain-google-genai==2.1.2
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.23
lmnr==0.5.0
markdownify==0.14.1
monotonic==1.6
ollama==0.4.7
openai==1.70.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.39.0
opentelemetry-instrumentation-anthropic==0.39.0
opentelemetry-instrumentation-bedrock==0.39.0
opentelemetry-instrumentation-chromadb==0.39.0
opentelemetry-instrumentation-cohere==0.39.0
opentelemetry-instrumentation-google-generativeai==0.39.0
opentelemetry-instrumentation-groq==0.39.0
opentelemetry-instrumentation-haystack==0.39.0
opentelemetry-instrumentation-lancedb==0.39.0
opentelemetry-instrumentation-langchain==0.39.0
opentelemetry-instrumentation-llamaindex==0.39.0
opentelemetry-instrumentation-marqo==0.39.0
opentelemetry-instrumentation-milvus==0.39.0
opentelemetry-instrumentation-mistralai==0.39.0
opentelemetry-instrumentation-ollama==0.39.0
opentelemetry-instrumentation-openai==0.39.0
opentelemetry-instrumentation-pinecone==0.39.0
opentelemetry-instrumentation-qdrant==0.39.0
opentelemetry-instrumentation-replicate==0.39.0
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.39.0
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.39.0
opentelemetry-instrumentation-transformers==0.39.0
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.39.0
opentelemetry-instrumentation-watsonx==0.39.0
opentelemetry-instrumentation-weaviate==0.39.0
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.3
opentelemetry-util-http==0.52b1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
playwright==1.51.0
posthog==3.23.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.1
pydantic-core==2.33.0
pyee==12.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing-extensions==4.13.0
typing-inspection==0.4.0
urllib3==2.3.0
wrapt==1.17.2
zipp==3.21.0
zstandard==0.23.0
